requirements.txt
Pillow==9.5.0
pytest==7.4.2
